pipeline_service/requirements.txt
dlt
dlt[postgres]
fastapi
httpx
pydantic
python-dotenv
sqlalchemy
uvicorn